Don't fail if only PROMETHEUS_PORT is set
