+++ pkg/ebpf/httpfltr/httpfltr_darwin.go
@@ -1,0 +1,6 @@
+package httpfltr
+
+func findNamespace(_ int32) (uint32, error) {
+	// convenience method to allow unit tests compiling in Darwin
+	return 1, nil
+}
